fixed loading of game library on x64

+++ neo/framework/Common.cpp
@@ -196,7 +196,7 @@ private:
 	idStrList					warningList;
 	idStrList					errorList;
 
-	int							gameDLL;
+	intptr_t					gameDLL;
 
 	idLangDict					languageDict;
